pkg/kubelet: move the capabilities related code to util.go
